Kyra's most recent PR made irrelevant most of my changes.
I kept a few of them to show off, but I will happily delete
them once I demonstrate what I've implemented and how.
Draper is pretty neat, but so many of the conditionals are in the
nitty gritty of the rails helpers, so when using JavaScript
it doesn't even help.